remove okteto init
