+++ cmd/utils/stack.go
@@ -16,11 +16,9 @@ package utils
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	"github.com/okteto/okteto/pkg/discovery"
 	"github.com/okteto/okteto/pkg/filesystem"
-	"github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
 )
 
@@ -57,27 +55,4 @@ func LoadStackContext(stackPaths []string) (*model.ContextResource, error) {
 		}
 	}
 	return ctxResource, nil
-}
-
-// GetStackFiles returns the list of stack files on a path
-func GetStackFiles(cwd string) []string {
-	result := []string{}
-	paths, err := os.ReadDir(cwd)
-	if err != nil {
-		return nil
-	}
-	for _, info := range paths {
-		if info.IsDir() {
-			continue
-		}
-		if strings.HasPrefix(info.Name(), "docker-compose") || strings.HasPrefix(info.Name(), "okteto-compose") || strings.HasPrefix(info.Name(), "okteto-stack") || strings.HasPrefix(info.Name(), "stack") {
-			result = append(result, info.Name())
-		}
-	}
-
-	if err != nil {
-		log.Infof("could not get stack files: %s", err.Error())
-	}
-	return result
-
 }
